Atualiza controle de acesso e estrutura do projeto
* Adiciona referências de pacotes e pasta para migrações.
* Implementa novos métodos no `AulaController` para recuperação de aulas por educador e usuário autenticado.
* Atualiza `InscricaoController` para autorização de alunos em inscrições.
* Melhora `UsuarioController` com métodos para obter e atualizar informações do usuário autenticado.
* Remove `EducadorId` do modelo `AulaRequest`, passando-o como argumento.
* Adiciona extensões para `ClaimsPrincipal` para facilitar a recuperação do ID do usuário.
* Atualiza classes de mapeamento e serviços para refletir novas lógicas de negócios.
* Modifica e cria migrações para novas tabelas e colunas no banco de dados.
* Atualiza sistema de autenticação para incluir tipo de reivindicação de função.

+++ Controllers/InscricaoController.cs
@@ -1,4 +1,6 @@
+using AlfabetizaFeso.Api.Extensions;
 using AlfabetizaFeso.Api.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AlfabetizaFeso.Api.Controllers;
@@ -15,9 +17,13 @@ public class InscricaoController : ControllerBase
     }
 
 
-    [HttpPost("aluno/{alunoId}/aula/{aulaId}")]
-    public async Task<IActionResult> Inscrever(int alunoId, int aulaId)
+    [HttpPost("aula/{aulaId}")]
+    [Authorize(Roles = "aluno")]
+    public async Task<IActionResult> Inscrever(int aulaId)
     {
+        if (User.GetUserId() is not int alunoId)
+            return Unauthorized();
+
         try
         {
             var inscricao = await _inscricaoService.AdicionarAsync(alunoId, aulaId);
@@ -30,9 +36,13 @@ public class InscricaoController : ControllerBase
     }
 
 
-    [HttpDelete("aluno/{alunoId}/aula/{aulaId}")]
-    public async Task<IActionResult> Cancelar(int alunoId, int aulaId)
+    [HttpDelete("aula/{aulaId}")]
+    [Authorize(Roles = "aluno")]
+    public async Task<IActionResult> Cancelar(int aulaId)
     {
+        if (User.GetUserId() is not int alunoId)
+            return Unauthorized();
+
         var removido = await _inscricaoService.RemoverAsync(alunoId, aulaId);
         if (!removido)
             return NotFound();
@@ -51,6 +61,16 @@ public class InscricaoController : ControllerBase
     [HttpGet("aluno/{alunoId}/aulas")]
     public async Task<IActionResult> ListarAulasPorAluno(int alunoId)
     {
+        var aulas = await _inscricaoService.ObterAulasInscritasPorAlunoIdAsync(alunoId);
+        return Ok(aulas);
+    }
+
+    [HttpGet("minhas-inscricoes")]
+    public async Task<IActionResult> ListarMinhasAulas()
+    {
+        if (User.GetUserId() is not int alunoId)
+            return Unauthorized();
+
         var aulas = await _inscricaoService.ObterAulasInscritasPorAlunoIdAsync(alunoId);
         return Ok(aulas);
     }
